// frontend/src/types/generated/AgentNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApiKeyConfig } from "./ApiKeyConfig";

export type AgentNode = { model: string, prompt: string, temperature: number, api_key_config: ApiKeyConfig | null, tools: Array<string> | null, };